+++ src/app/login/page.tsx
@@ -1,219 +1,4 @@
-'use client';
-
-import { useState, Suspense } from 'react';
-import { signIn } from 'next-auth/react';
-import { useRouter, useSearchParams } from 'next/navigation';
-
-// Wrap the main content in Suspense to handle searchParams in Client Components safely
+import LoginForm from '@/components/login/login-form';
 export default function LoginPage() {
-  return (
-    <Suspense
-      fallback={
-        <div className="flex h-screen items-center justify-center">
-          Loading...
-        </div>
-      }
-    >
-      <LoginForm />
-    </Suspense>
-  );
-}
-
-function LoginForm() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-
-  // State for form inputs
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  // Check for NextAuth errors in URL (e.g., ?error=CredentialsSignin)
-  const authError = searchParams.get('error');
-
-  // ---------------------------------------------------------
-  // 1. Social Login Logic (Manual Redirects)
-  // ---------------------------------------------------------
-  const handleSocialLogin = (provider: 'kakao' | 'google') => {
-    setIsLoading(true);
-
-    // Exact URLs we discussed previously
-    const callbackUrl = `${window.location.origin}/auth/callback/${provider}`;
-
-    let authUrl = '';
-
-    if (provider === 'kakao') {
-      const CLIENT_ID = process.env.NEXT_PUBLIC_AUTH_KAKAO_ID;
-      authUrl = `https://kauth.kakao.com/oauth/authorize?client_id=${CLIENT_ID}&redirect_uri=${callbackUrl}&response_type=code`;
-    } else if (provider === 'google') {
-      const CLIENT_ID = process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID;
-      const SCOPE = 'openid email profile';
-      authUrl = `https://accounts.google.com/o/oauth2/v2/auth?client_id=${CLIENT_ID}&redirect_uri=${callbackUrl}&response_type=code&scope=${SCOPE}`;
-    }
-
-    // Redirect the user
-    window.location.href = authUrl;
-  };
-
-  // ---------------------------------------------------------
-  // 2. Credentials Login Logic (Email/Password)
-  // ---------------------------------------------------------
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError('');
-
-    try {
-      // Calls the 'credentials-login' provider in your auth.ts
-      const result = await signIn('credentials-login', {
-        email,
-        password,
-        redirect: false, // We handle redirect manually to show errors nicely
-      });
-
-      if (result?.error) {
-        setError('Invalid email or password.');
-      } else {
-        router.push('/dashboard'); // Success! Redirect to app
-        router.refresh(); // Refresh router to update session state
-      }
-    } catch (err) {
-      setError('Something went wrong. Please try again.');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-50 px-4 py-12 sm:px-6 lg:px-8">
-      <div className="w-full max-w-md space-y-8 rounded-xl bg-white p-10 shadow-lg">
-        {/* Header */}
-        <div className="text-center">
-          <h2 className="mt-2 text-3xl font-bold tracking-tight text-gray-900">
-            Welcome back
-          </h2>
-          <p className="mt-2 text-sm text-gray-600">
-            Please sign in to your account
-          </p>
-        </div>
-
-        {/* Error Message Display */}
-        {(error || authError) && (
-          <div className="rounded-md bg-red-50 p-4 text-center text-sm text-red-500">
-            {error || 'Authentication failed. Please try again.'}
-          </div>
-        )}
-
-        {/* Social Buttons Section */}
-        <div className="mt-8 space-y-4">
-          <button
-            onClick={() => handleSocialLogin('kakao')}
-            disabled={isLoading}
-            className="flex w-full items-center justify-center gap-3 rounded-lg bg-[#FEE500] px-4 py-3 text-sm font-semibold text-black shadow-sm transition-colors hover:bg-[#FDD835] focus:outline-none disabled:opacity-50"
-          >
-            {/* Kakao SVG Icon */}
-            <svg viewBox="0 0 24 24" className="h-5 w-5 fill-current">
-              <path d="M12 3C5.373 3 0 6.697 0 11.258c0 2.925 2.228 5.494 5.61 6.836-.254.936-1.077 3.822-1.22 4.414-.142.593.24.606.52.41 1.08-.76 4.755-3.23 5.474-3.722.533.076 1.08.115 1.638.115 6.627 0 12-3.697 12-8.258C24 6.697 18.627 3 12 3z" />
-            </svg>
-            Sign in with Kakao
-          </button>
-
-          <button
-            onClick={() => handleSocialLogin('google')}
-            disabled={isLoading}
-            className="flex w-full items-center justify-center gap-3 rounded-lg border border-gray-300 bg-white px-4 py-3 text-sm font-semibold text-gray-700 shadow-sm transition-colors hover:bg-gray-50 focus:outline-none disabled:opacity-50"
-          >
-            {/* Google SVG Icon */}
-            <svg viewBox="0 0 24 24" className="h-5 w-5">
-              <path
-                d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-                fill="#4285F4"
-              />
-              <path
-                d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                fill="#34A853"
-              />
-              <path
-                d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.84z"
-                fill="#FBBC05"
-              />
-              <path
-                d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                fill="#EA4335"
-              />
-            </svg>
-            Sign in with Google
-          </button>
-        </div>
-
-        {/* Divider */}
-        <div className="relative">
-          <div className="absolute inset-0 flex items-center">
-            <div className="w-full border-t border-gray-300" />
-          </div>
-          <div className="relative flex justify-center text-sm">
-            <span className="bg-white px-2 text-gray-500">
-              Or continue with
-            </span>
-          </div>
-        </div>
-
-        {/* Credentials Form */}
-        <form className="space-y-6" onSubmit={handleSubmit}>
-          <div>
-            <label
-              htmlFor="email"
-              className="block text-sm font-medium text-gray-700"
-            >
-              Email address
-            </label>
-            <div className="mt-1">
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                className="block w-full rounded-md border border-gray-300 px-3 py-2 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 focus:outline-none sm:text-sm"
-              />
-            </div>
-          </div>
-
-          <div>
-            <label
-              htmlFor="password"
-              className="block text-sm font-medium text-gray-700"
-            >
-              Password
-            </label>
-            <div className="mt-1">
-              <input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete="current-password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                className="block w-full rounded-md border border-gray-300 px-3 py-2 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 focus:outline-none sm:text-sm"
-              />
-            </div>
-          </div>
-
-          <div>
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="flex w-full justify-center rounded-md border border-transparent bg-indigo-600 px-4 py-2 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 focus:outline-none disabled:bg-indigo-400"
-            >
-              {isLoading ? 'Signing in...' : 'Sign in'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
+  return <LoginForm />;
 }
